refactor: migrated /api/tags
