+++ sites/geohub/static/api/swagger/spec.yaml
@@ -2572,139 +2572,6 @@ paths:
         This endpoint is to create a feature JSON object for selected STAC item asset. This endpoint only can work for catalog type of STAC. STAC API item should use other endpoint of `/stac/{type}/{collection}/{...item}/{asset}.
 
         If multiple `url`s are specified, the endpoint will create a mosaicjson feature.
-  /tags:
-    get:
-      summary: Tag search API
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-              examples:
-                Example for tags?key=extent:
-                  value:
-                    extent:
-                      - key: extent
-                        value: Afghanistan
-                        count: 2
-                      - key: extent
-                        value: Africa
-                        count: 5
-                      - key: extent
-                        value: America
-                        count: 1
-                      - key: extent
-                        value: Asia
-                        count: 5
-                      - key: extent
-                        value: Burundi
-                        count: 3
-                      - key: extent
-                        value: Cambodia
-                        count: 2
-                      - key: extent
-                        value: Conflict areas
-                        count: 1
-                      - key: extent
-                        value: Ethiopia
-                        count: 3
-                      - key: extent
-                        value: EU
-                        count: 3
-                      - key: extent
-                        value: Europe
-                        count: 9
-                      - key: extent
-                        value: Ghana
-                        count: 2
-                      - key: extent
-                        value: Global
-                        count: 3845
-                      - key: extent
-                        value: Greenland and Antarctica
-                        count: 1
-                      - key: extent
-                        value: India
-                        count: 2
-                      - key: extent
-                        value: Kenya
-                        count: 10
-                      - key: extent
-                        value: Madagascar
-                        count: 2
-                      - key: extent
-                        value: Malawi
-                        count: 4
-                      - key: extent
-                        value: Mali
-                        count: 2
-                      - key: extent
-                        value: Mozambique
-                        count: 2
-                      - key: extent
-                        value: Nepal
-                        count: 2
-                      - key: extent
-                        value: Nigeria
-                        count: 12
-                      - key: extent
-                        value: North America
-                        count: 1
-                      - key: extent
-                        value: Oceania
-                        count: 1
-                      - key: extent
-                        value: OECD
-                        count: 7
-                      - key: extent
-                        value: Pakistan
-                        count: 6
-                      - key: extent
-                        value: Rwanda
-                        count: 6
-                      - key: extent
-                        value: South America
-                        count: 1
-                      - key: extent
-                        value: SouthEast Asia
-                        count: 2
-                      - key: extent
-                        value: Tanzania
-                        count: 8
-                      - key: extent
-                        value: Uganda
-                        count: 10
-                      - key: extent
-                        value: West Africa
-                        count: 1
-                      - key: extent
-                        value: Zambia
-                        count: 2
-      operationId: get-tags
-      description: 'This API can return the list of available tags. If you simply query by /tags, it returns all tags. You can also filter a key value to list available tags.'
-      parameters:
-        - schema:
-            type: string
-            enum:
-              - extent
-              - granularity
-              - keyword
-              - layertype
-              - resolution
-              - schema
-              - sdg_goal
-              - sdg_target
-              - stac
-              - theme
-              - type
-              - year
-          in: query
-          name: key
-          description: Optional parameter to filter by key
-      tags:
-        - tags
   /vector/statistics:
     get:
       summary: Vector tile statistics API
